Use 'administrator' instead of 'operator'
While the term "operator" is sometimes used to mean "person maintaining a
Kubernetes cluster", this usage is rare and can easily be confused with
"software extension to Kubernetes".

To reduce this ambiguity and better align vocabulary with community
usage -- e.g., Certified Kubernetes Administrator -- we will use
administrator from now on.