Fix include path
mkdocs-include-markdown-plugin recently changed the way it resolves
paths. Specifically:

"""
- Local files:
  - Absolute paths (starting with a path separator).
  - Relative from the file that includes them (starting with `./` or `../`).
  - Relative from the _docs/_ directory.
"""

Hence a path like `_common.include` is now interpreted as relative to
the _docs/_ directory.

This commit adds `./` in front of all paths, so that they are correctly
interpreted as relative to the file where the include macro is used.

Further reading:
- https://github.com/mondeja/mkdocs-include-markdown-plugin/commit/eb9f6ed90d8da15bf2e23b9b88dced2d8f5391f6#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5