Avoid some redirects
linkchecker changed its default behavior to warn if a link points to a
redirect, i.e., HTTP 30x. [1]. As a result, we got 61 warnings during
external link checking. [2]

Instead of disabling the warnings, I decided to go ahead and
incrementally fix them. This PR represents the first increment.

Most of those warnings are "forgot to put a slash at the end" warnings.

[1]: https://github.com/linkchecker/linkchecker/commit/beaf9399f88c1d2e6c13f8ff14f54a304eda3c91
[2]: https://github.com/elastisys/compliantkubernetes/actions/runs/6233244612/job/16918039321