Unify linkchecker config files (#923)
We totally want to check not only URLs, but also anchors, for better visitor
experience. Amongst others, we use this to avoid breaking anchors in the ToS.
Unfortunately, Some pages contain dynamically generated anchors, which fail
AnchorCheck. We previously ran linkchecker twice, with two ignore lists.

This approach led to a few issues:

1. We forgot to update both config files (without AnchorCheck and with
AnchorCheck) at the same time, which led to many false positive.
2. Linkchecking takes twice the amount of time.

To avoid these issues, we decided to only run linkchecker once, with
AnchorCheck. This means that URLs to some pages won't be checked at all.
We accept this risk: We prefer to slightly underalert, as opposed to overalert.