Rank up application developer pages
We recently added configuration reference. Those pages are long and detailed
and tend to come up when searching for everything. This may confuse application
developers.

Therefore, we decided to boost the ranking of pages relevant for application
developers. We also decided to make sure that searching for "Alertmanager"
bring the most relevant application developer page, by adding
"via Alertmanager" in the title.