Refresh list of ignored errors
Seems like we got completely wrong how to produce the list of ignored errors. This means that those errors were not
ignored and we did end up being someone else's uptime monitor.

This commit rectified this situation.

This commit is minor, because it does not materially affect the elastisys.io website and is only meant to fix the
implementation of an existing QA requirement.